+++ src/main/java/com/example/demo/services/TicketService.java
@@ -5,7 +5,7 @@ import com.example.demo.entity.Ticket;
 import java.util.List;
 
 public interface TicketService {
-    Ticket getById (int id);
+    Ticket getById (Long id);
 
     List<Ticket> getAllTickets();
